+++ packages/prover/README.md
@@ -11,6 +11,27 @@ Typescript REST client for the [Ethereum Consensus API spec](https://github.com/
 
 ## Usage
 
+You can use the `@lodestar/prover` in two ways, as a Web3 Provider and as proxy. For prover use case see below example.
+
+```ts
+import Web3 from "web3";
+import {makeProvableProvider, LightNode} from "@lodestar/prover";
+
+const provider = new Web3.providers.HttpProvider("https://lodestar-sepoliarpc.chainsafe.io");
+const proveableProvider = makeProvableProvider(provider, {
+  mode: LightNode.Rest,
+  urls: ["https://lodestar-sepolia.chainsafe.io"],
+  network: "sepolia",
+});
+
+await proveableProvider.rootProvider.sync();
+
+const web3 = new Web3(proveableProvider);
+
+const address = "0xf97e180c050e5Ab072211Ad2C213Eb5AEE4DF134";
+const balance = await web3.eth.getBalance(address, "latest");
+console.log({balance, address});
+```
 
 ## Prerequisites
