+++ packages/prover/src/constants.ts
@@ -1,3 +1,4 @@
 // https://github.com/ethereum/consensus-specs/blob/dev/specs/altair/light-client/p2p-interface.md#configuration
 export const MAX_REQUEST_LIGHT_CLIENT_UPDATES = 128;
 export const MAX_PAYLOAD_HISTORY = 32;
+export const UNVERIFIED_RESPONSE_CODE = -33091;
