+++ packages/prover/src/verified_requests/eth_getBalance.ts
@@ -1,0 +1,30 @@
+import {ELVerifiedRequestHandler} from "../interfaces.js";
+import {hexToBuffer, numberToHex} from "../utils/conversion.js";
+import {getELProof} from "../utils/execution.js";
+import {generateRPCResponseForPayload, generateUnverifiedResponseForPayload} from "../utils/json_rpc.js";
+import {isValidAccount, isValidStorageKeys} from "../utils/verification.js";
+
+export const ethGetBalance: ELVerifiedRequestHandler<[address: string, block?: number | string], string> = async ({
+  handler,
+  payload,
+  rootProvider,
+}) => {
+  const {
+    params: [address, block],
+  } = payload;
+  const executionPayload = rootProvider.getExecutionPayload(block ?? "latest");
+  const proof = await getELProof(handler, [address, [], numberToHex(executionPayload.blockNumber)]);
+
+  if (
+    (await isValidAccount({
+      address: hexToBuffer(address),
+      stateRoot: executionPayload.stateRoot,
+      proof,
+    })) &&
+    (await isValidStorageKeys({storageKeys: [], proof}))
+  ) {
+    return generateRPCResponseForPayload(payload, proof.balance);
+  }
+
+  return generateUnverifiedResponseForPayload(payload, "eth_getBalance request can not be verified.");
+};
